+++ src/tests/App.test.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import React from 'react';
+import App from '../components/App';
 
 test('renders learn react link', () => {
   render(<App />);
