+++ src/utils/constants.ts
@@ -1,0 +1,7 @@
+import { ethers } from "ethers";
+
+export const REFUNDER_ADDRESS = "0x2d8cE02dd1644A9238e08430CaeA15a609503140";
+export const WETH_ADDRESS = "0xc778417E063141139Fce010982780140Aa0cD5Ab";
+export const MIN_FUNDS = ethers.utils.parseEther("0.001")
+
+export const SERVICE_URL = `https://yacate-relay-rinkeby.herokuapp.com/v1/transactions/execute/generic`
